requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
mfrc522==0.0.7
requests==2.28.1
RPi.GPIO==0.7.1
spidev==3.6
urllib3==1.26.12
